requirements.txt
pyserial>=3.5
tqdm>=4.65.0